Lint packages.
