+++ packages/gasket-engine/test/exec-apply-sync.test.js
@@ -55,7 +55,7 @@ describe('The execApplySync method', () => {
       };
     });
 
-    const PluginEngine = require('..');
+    const PluginEngine = require('../lib/engine');
     engine = new PluginEngine(mockConfig);
   });
