+++ packages/gasket-plugin-command/package.json
@@ -41,9 +41,11 @@
   },
   "devDependencies": {
     "@godaddy/dmd": "^1.0.0",
+    "@oclif/command": "^1.8.16",
     "@oclif/config": "^1.15.0",
     "@oclif/errors": "^1.1.2",
     "assume": "^2.3.0",
+    "assume-sinon": "^1.1.0",
     "eslint": "^7.17.0",
     "eslint-config-godaddy": "^4.0.1",
     "eslint-plugin-jest": "^24.1.3",
@@ -53,6 +55,7 @@
     "jsdoc-to-markdown": "^5.0.1",
     "mocha": "^8.2.1",
     "nyc": "^15.1.0",
+    "proxyquire": "^2.1.3",
     "sinon": "^9.2.3"
   },
   "peerDependencies": {
