+++ packages/gasket-plugin-config/package.json
@@ -44,7 +44,8 @@
     "eslint-plugin-json": "^2.1.2",
     "eslint-plugin-mocha": "^8.0.0",
     "eslint-plugin-unicorn": "^39.0.0",
-    "jest": "^26.6.3"
+    "jest": "^26.6.3",
+    "sinon": "^12.0.1"
   },
   "eslintConfig": {
     "extends": [
