+++ packages/gasket-plugin-docs-graphs/package.json
@@ -33,6 +33,7 @@
   },
   "homepage": "https://github.com/godaddy/gasket/tree/master/packages/gasket-plugin-doc-graph",
   "devDependencies": {
+    "assume": "^2.3.0",
     "eslint": "^7.17.0",
     "eslint-config-godaddy": "^4.0.1",
     "eslint-plugin-json": "^2.1.2",
