+++ packages/gasket-plugin-manifest/package.json
@@ -35,7 +35,9 @@
   },
   "homepage": "https://github.com/godaddy/gasket/tree/master/packages/gasket-plugin-manifest#readme",
   "dependencies": {
-    "escape-string-regexp": "^1.0.5"
+    "deepmerge": "^4.2.2",
+    "escape-string-regexp": "^1.0.5",
+    "mkdirp": "^1.0.4"
   },
   "devDependencies": {
     "assume": "^2.3.0",
@@ -46,6 +48,7 @@
     "eslint-plugin-unicorn": "^39.0.0",
     "mocha": "^8.2.1",
     "nyc": "^15.1.0",
+    "proxyquire": "^2.1.3",
     "sinon": "^9.2.3"
   },
   "eslintConfig": {
