+++ packages/gasket-plugin-metrics/package.json
@@ -40,6 +40,7 @@
   },
   "devDependencies": {
     "assume": "^2.3.0",
+    "assume-sinon": "^1.1.0",
     "eslint": "^7.17.0",
     "eslint-config-godaddy": "^4.0.1",
     "eslint-plugin-json": "^2.1.2",
