+++ packages/gasket-plugin-start/package.json
@@ -39,6 +39,7 @@
   "homepage": "https://github.com/godaddy/gasket/tree/master/packages/gasket-plugin-start",
   "devDependencies": {
     "assume": "^2.3.0",
+    "assume-sinon": "^1.1.0",
     "eslint": "^7.17.0",
     "eslint-config-godaddy": "^4.0.1",
     "eslint-plugin-json": "^2.1.2",
