+++ packages/gasket-plugin-swagger/package.json
@@ -41,6 +41,9 @@
   },
   "devDependencies": {
     "@babel/core": "^7.12.10",
+    "@babel/plugin-transform-runtime": "^7.16.5",
+    "@babel/preset-env": "^7.16.5",
+    "@babel/preset-react": "^7.16.5",
     "@babel/register": "^7.12.10",
     "assume": "^2.3.0",
     "assume-sinon": "^1.1.0",
