+++ packages/gasket-redux/package.json
@@ -51,6 +51,7 @@
     "babel-eslint": "^10.1.0",
     "babel-jest": "^23.2.0",
     "eslint": "^7.17.0",
+    "eslint-config-godaddy": "^5.0.1",
     "eslint-plugin-jest": "^24.1.3",
     "eslint-plugin-json": "^2.1.2",
     "eslint-plugin-jsx-a11y": "^6.2.1",
